3:52 7/3/2017

+++ local/database/migrations/2017_02_27_075216_create_chapters_table.php
@@ -15,7 +15,7 @@ class CreateChaptersTable extends Migration
     {
         Schema::create('chapters', function (Blueprint $table) {
             $table->increments('chapter_id');
-            $table->integer('chapter_serial')->default(0);
+            $table->float('chapter_serial')->default(0);
             $table->text('chapter_title');
             $table->text('chapter_content');
             $table->text('chapter_status');
